requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tiktoken>=0.5.0
rich>=13.0.0
pyyaml>=6.0.0
aiohttp>=3.9.0
